Completed setup command

+++ bin/commands/internal/setup.sh
@@ -4,17 +4,27 @@ SCRIPT_DIRECTORY="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pw
 SETUP_FILE="${SCRIPT_DIRECTORY}/../../include/setup_options.sh"
 CONTENT_HUB_FARM_DIRECTORY=`dirname "$(dirname "$(dirname "$SCRIPT_DIRECTORY")"))"`
 
-# @TODO: Check that docker-compose.yml exists, if so... ask to do a quick-setup.
-
-# @TODO: Check if there are containers running, if so ask to destroy them, including volumes.
-
-# @TODO: Check if there is codebase in the 'html' directory. If so, ask to delete it.
-
-# @TODO: Add default values to configuration variables.
-
 echo "Configuration Setup for Content Hub Farm"
 echo "----------------------------------------"
 echo ""
+
+# If provided options --fast, then recreate docker-compose.yml from the saved configuration.
+if [[ $1 == "--fast" && -f "$SETUP_FILE" ]] ; then
+  echo "Regenerating docker-compose.yml and ngrok.yml from previously saved configuration."
+  # Regenerating docker-compose.yml from setup_options.sh.
+  source $SETUP_FILE
+  TEMPLATE='default'
+  if [ "${CONFIG_ENABLE_NFS}" == 1 ] ; then
+    TEMPLATE='nfs'
+  fi
+  echo "Regenerating docker-composer.yml..."
+  sh ${CONTENT_HUB_FARM_DIRECTORY}/bin/templates/docker-compose.sh $TEMPLATE
+  # Creating ngrok.yml.
+  echo "Regenerating ~/.ngrok2/ngrok.yml..."
+  sh ${CONTENT_HUB_FARM_DIRECTORY}/bin/templates/ngrok.sh
+  echo "Finished Setup."
+  exit
+fi
 
 # Start Questionaire.
 echo "#!/bin/bash" > ${SETUP_FILE};
@@ -72,12 +82,41 @@ done
 
 # Volume Device Path.
 echo "# Volume basepath and Ngrok Token." >> ${SETUP_FILE}
-echo "CONFIG_VOLUME_DEVICE_PATH=$CONTENT_HUB_FARM_DIRECTORY" >> ${SETUP_FILE}
+echo "CONFIG_VOLUME_DEVICE_PATH=\"$CONTENT_HUB_FARM_DIRECTORY\"" >> ${SETUP_FILE}
 
 # Ngrok Token.
 echo "Please Provide your Ngrok Token. You can obtain it from https://dashboard.ngrok.com/auth."
 read -p "Ngrok token: " CONFIG_NGROK_TOKEN
 echo "CONFIG_NGROK_TOKEN=\"${CONFIG_NGROK_TOKEN}\";" >> ${SETUP_FILE}
+echo "" >> ${SETUP_FILE}
+echo ""
+
+# Build code using public or private Content Hub repository.
+echo "Build Site Codebase."
+echo "Do you want to install Acquia Content Hub from Public or Private Repository?"
+CONFIG_BUILD_CODE_SOURCE="public"
+CONFIG_BUILD_CODE_BRANCH="^2"
+options=("Public" "Private")
+select opt in "${options[@]}"
+do
+  case $opt in
+    "Public")
+      echo "Using latest public release of Content Hub 2.x from Drupal.org."
+      break;
+      ;;
+    "Private")
+      CONFIG_BUILD_CODE_SOURCE="private"
+      echo "Using Acquia's Private repository."
+      read -p "Insert a Branch (8.x-2.x): " CONFIG_BUILD_CODE_BRANCH
+      CONFIG_BUILD_CODE_BRANCH="${CONFIG_BUILD_CODE_BRANCH:-"8.x-2.x"}"
+      break;
+      ;;
+    *) echo "Invalid option $REPLY";;
+  esac
+done
+echo "# Build Site Codebase." >> ${SETUP_FILE}
+echo "CONFIG_BUILD_CODE_SOURCE=\"${CONFIG_BUILD_CODE_SOURCE}\";" >> ${SETUP_FILE}
+echo "CONFIG_BUILD_CODE_BRANCH=\"${CONFIG_BUILD_CODE_BRANCH}\";" >> ${SETUP_FILE}
 echo ""
 
 echo "Configuration Options saved in './bin/include/setup_options.sh'."
@@ -92,6 +131,5 @@ sh ${CONTENT_HUB_FARM_DIRECTORY}/bin/templates/docker-compose.sh $TEMPLATE
 
 # Creating ngrok.yml.
 sh ${CONTENT_HUB_FARM_DIRECTORY}/bin/templates/ngrok.sh
-
 
 echo "Finished Setup."
